+++ src/restruct_poc/model/model.hpp
@@ -25,39 +25,42 @@ struct Model {
         const Array_6& velocity = Array_6{0., 0., 0., 0., 0., 0.},
         const Array_6& acceleration = Array_6{0., 0., 0., 0., 0., 0.}
     ) {
-        this->nodes.push_back(Node(nodes.size(), position, displacement, velocity, acceleration));
+        this->nodes.push_back(
+            Node(static_cast<int>(nodes.size()), position, displacement, velocity, acceleration)
+        );
         return this->nodes.back();
     }
 
     /// Adds a rigid constraint to the model and returns the constraint.
     Constraint AddRigidConstraint(const Node& node1, const Node& node2) {
-        this->constraints.push_back(
-            Constraint(ConstraintType::kRigid, this->constraints.size(), node1, node2)
-        );
+        this->constraints.push_back(Constraint(
+            ConstraintType::kRigid, static_cast<int>(this->constraints.size()), node1, node2
+        ));
         return this->constraints.back();
     }
 
     /// Adds a prescribed boundary condition constraint to the model and returns the constraint.
     Constraint AddPrescribedBC(const Node& node, const Array_3& ref_position = {0., 0., 0.}) {
         this->constraints.push_back(Constraint(
-            ConstraintType::kPrescribedBC, this->constraints.size(), InvalidNode, node, ref_position
+            ConstraintType::kPrescribedBC, static_cast<int>(this->constraints.size()), InvalidNode,
+            node, ref_position
         ));
         return this->constraints.back();
     }
 
     /// Adds a fixed boundary condition constraint to the model and returns the constraint.
     Constraint AddFixedBC(const Node& node) {
-        this->constraints.push_back(
-            Constraint(ConstraintType::kFixedBC, this->constraints.size(), InvalidNode, node)
-        );
+        this->constraints.push_back(Constraint(
+            ConstraintType::kFixedBC, static_cast<int>(this->constraints.size()), InvalidNode, node
+        ));
         return this->constraints.back();
     }
 
     /// Adds a cylindrical constraint to the model and returns the constraint.
     Constraint AddCylindricalConstraint(const Node& node1, const Node& node2) {
-        this->constraints.push_back(
-            Constraint(ConstraintType::kCylindrical, this->constraints.size(), node1, node2)
-        );
+        this->constraints.push_back(Constraint(
+            ConstraintType::kCylindrical, static_cast<int>(this->constraints.size()), node1, node2
+        ));
         return this->constraints.back();
     }
 
@@ -66,13 +69,14 @@ struct Model {
         const Node& node1, const Node& node2, const Array_3& axis, float* control
     ) {
         this->constraints.push_back(Constraint(
-            ConstraintType::kRotationControl, this->constraints.size(), node1, node2, axis, control
+            ConstraintType::kRotationControl, static_cast<int>(this->constraints.size()), node1,
+            node2, axis, control
         ));
         return this->constraints.back();
     }
 
     /// Returns the number of nodes in the model
-    int NumNodes() { return this->nodes.size(); }
+    size_t NumNodes() { return this->nodes.size(); }
 };
 
 }  // namespace openturbine
