+++ src/restruct_poc/system/update_state.hpp
@@ -41,7 +41,7 @@ inline void UpdateState(Beams& beams, View_Nx7 Q, View_Nx6 V, View_Nx6 A) {
         }
     );
 
-    auto range_policy = Kokkos::TeamPolicy<>(beams.num_elems, Kokkos::AUTO());
+    auto range_policy = Kokkos::TeamPolicy<>(static_cast<int>(beams.num_elems), Kokkos::AUTO());
 
     Kokkos::parallel_for(
         "InterpolateToQuadraturePoints", range_policy,
